+++ cli/k8s_client/k8s_client.go
@@ -23,7 +23,6 @@ import (
 	"k8s.io/api/policy/v1beta1"
 	v13 "k8s.io/api/rbac/v1"
 	storagev1 "k8s.io/api/storage/v1"
-	v1beta12 "k8s.io/api/storage/v1beta1"
 	apiextensionv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apiextension "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
@@ -1657,22 +1656,6 @@ func (k *KubeClient) PatchClusterRoleBindingByLabel(label string, patchBytes []b
 	return nil
 }
 
-// GetBetaCSIDriverByLabel returns a CSI driver object matching the specified label if it is unique
-func (k *KubeClient) GetBetaCSIDriverByLabel(label string) (*v1beta12.CSIDriver, error) {
-	CSIDrivers, err := k.GetBetaCSIDriversByLabel(label)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(CSIDrivers) == 1 {
-		return &CSIDrivers[0], nil
-	} else if len(CSIDrivers) > 1 {
-		return nil, fmt.Errorf("multiple CSI drivers have the label %s", label)
-	} else {
-		return nil, utils.NotFoundError(fmt.Sprintf("no CSI drivers have the label %s", label))
-	}
-}
-
 // GetCSIDriverByLabel returns a CSI driver object matching the specified label if it is unique
 func (k *KubeClient) GetCSIDriverByLabel(label string) (*storagev1.CSIDriver, error) {
 	CSIDrivers, err := k.GetCSIDriversByLabel(label)
@@ -1689,21 +1672,6 @@ func (k *KubeClient) GetCSIDriverByLabel(label string) (*storagev1.CSIDriver, er
 	}
 }
 
-// GetBetaCSIDriversByLabel returns all CSI driver objects matching the specified label
-func (k *KubeClient) GetBetaCSIDriversByLabel(label string) ([]v1beta12.CSIDriver, error) {
-	listOptions, err := k.listOptionsFromLabel(label)
-	if err != nil {
-		return nil, err
-	}
-
-	CSIDriverList, err := k.clientset.StorageV1beta1().CSIDrivers().List(ctx(), listOptions)
-	if err != nil {
-		return nil, err
-	}
-
-	return CSIDriverList.Items, nil
-}
-
 // GetCSIDriversByLabel returns all CSI driver objects matching the specified label
 func (k *KubeClient) GetCSIDriversByLabel(label string) ([]storagev1.CSIDriver, error) {
 	listOptions, err := k.listOptionsFromLabel(label)
@@ -1717,24 +1685,6 @@ func (k *KubeClient) GetCSIDriversByLabel(label string) ([]storagev1.CSIDriver, 
 	}
 
 	return CSIDriverList.Items, nil
-}
-
-// CheckBetaCSIDriverExistsByLabel returns true if one or more CSI Driver objects
-// matching the specified label exist.
-func (k *KubeClient) CheckBetaCSIDriverExistsByLabel(label string) (bool, string, error) {
-	CSIDrivers, err := k.GetBetaCSIDriversByLabel(label)
-	if err != nil {
-		return false, "", err
-	}
-
-	switch len(CSIDrivers) {
-	case 0:
-		return false, "", nil
-	case 1:
-		return true, CSIDrivers[0].Namespace, nil
-	default:
-		return true, "<multiple>", nil
-	}
 }
 
 // CheckCSIDriverExistsByLabel returns true if one or more CSI Driver objects
@@ -1753,26 +1703,6 @@ func (k *KubeClient) CheckCSIDriverExistsByLabel(label string) (bool, string, er
 	default:
 		return true, "<multiple>", nil
 	}
-}
-
-// DeleteBetaCSIDriverByLabel deletes a CSI Driver object matching the specified label
-// in the namespace of the client.
-func (k *KubeClient) DeleteBetaCSIDriverByLabel(label string) error {
-	CSIDriver, err := k.GetBetaCSIDriverByLabel(label)
-	if err != nil {
-		return err
-	}
-
-	if err = k.clientset.StorageV1beta1().CSIDrivers().Delete(ctx(), CSIDriver.Name, k.deleteOptions()); err != nil {
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"label":       label,
-		"CSIDriverCR": CSIDriver.Name,
-	}).Debug("Deleted CSI driver.")
-
-	return nil
 }
 
 // DeleteCSIDriverByLabel deletes a CSI Driver object matching the specified label
@@ -1795,19 +1725,6 @@ func (k *KubeClient) DeleteCSIDriverByLabel(label string) error {
 	return nil
 }
 
-// DeleteBetaCSIDriver deletes a CSI Driver object matching the specified name
-func (k *KubeClient) DeleteBetaCSIDriver(name string) error {
-	if err := k.clientset.StorageV1beta1().CSIDrivers().Delete(ctx(), name, k.deleteOptions()); err != nil {
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"CSIDriverCR": name,
-	}).Debug("Deleted CSI driver.")
-
-	return nil
-}
-
 // DeleteCSIDriver deletes a CSI Driver object matching the specified name
 func (k *KubeClient) DeleteCSIDriver(name string) error {
 	if err := k.clientset.StorageV1().CSIDrivers().Delete(ctx(), name, k.deleteOptions()); err != nil {
@@ -1817,27 +1734,6 @@ func (k *KubeClient) DeleteCSIDriver(name string) error {
 	log.WithFields(log.Fields{
 		"CSIDriverCR": name,
 	}).Debug("Deleted CSI driver.")
-
-	return nil
-}
-
-// PatchBetaCSIDriverByLabel patches a deployment object matching the specified label
-// in the namespace of the client.
-func (k *KubeClient) PatchBetaCSIDriverByLabel(label string, patchBytes []byte, patchType types.PatchType) error {
-	csiDriver, err := k.GetBetaCSIDriverByLabel(label)
-	if err != nil {
-		return err
-	}
-
-	if _, err = k.clientset.StorageV1beta1().CSIDrivers().Patch(ctx(), csiDriver.Name,
-		patchType, patchBytes, patchOpts); err != nil {
-		return err
-	}
-
-	log.WithFields(log.Fields{
-		"label":      label,
-		"deployment": csiDriver.Name,
-	}).Debug("Patched Kubernetes CSI driver.")
 
 	return nil
 }